+++ src/toggleButton/index.unit.tsx
@@ -2,12 +2,6 @@ import React from 'react'
 import { shallow } from 'enzyme'
 
 import ToggleButton from 'toggleButton'
-import Item from '_utils/item/index'
-import Loader from 'loader'
-import CircleIcon from 'icon/circleIcon'
-import CheckIcon from 'icon/checkIcon'
-
-import { color } from '_utils/branding'
 
 const defaultProps = {
   name: 'foo',
@@ -46,12 +40,6 @@ describe('ToggleButton', () => {
       toggleButton = shallow(<ToggleButton {...defaultProps} checked />)
     })
 
-    it('Should render a CheckIcon', () => {
-      expect(toggleButton.find(Item).prop('rightAddon')).toEqual(
-        <CheckIcon iconColor={color.white} backgroundColor={color.primary} thin />,
-      )
-    })
-
     it('Should have the aria-pressed attribute to true', () => {
       const btn = toggleButton.find('button')
       expect(btn.prop('aria-pressed')).toBe(true)
@@ -61,12 +49,6 @@ describe('ToggleButton', () => {
   describe('unchecked state', () => {
     beforeEach(() => {
       toggleButton = shallow(<ToggleButton {...defaultProps} checked={false} />)
-    })
-
-    it('Should render a CircleIcon', () => {
-      expect(toggleButton.find(Item).prop('rightAddon')).toEqual(
-        <CircleIcon iconColor={color.primary} thin />,
-      )
     })
 
     it('Should have the aria-pressed attribute to false', () => {
@@ -80,10 +62,6 @@ describe('ToggleButton', () => {
       toggleButton = shallow(
         <ToggleButton {...defaultProps} status={ToggleButton.STATUS.LOADING} />,
       )
-    })
-
-    it('Should render a Loader', () => {
-      expect(toggleButton.find(Item).prop('rightAddon')).toEqual(<Loader size={24} inline />)
     })
 
     it('Should be disabled', () => {